microservices/order-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
marshmallow==3.20.1
flask-migrate==4.0.5
pika==1.3.2
requests==2.31.0
